fix graphics in minos / mino_stack, by factoring out the graphics section
